+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.6
 
+* Add `ComposeTraversable`. See #65.
 * Make the `Applicative` instance of `ComposeEither` more interesting by relying
   on the `Selective f` constraint. See #64.
 * Make the `Lift` instance lazier. See #63.